Re-factor some margin functions to support bigger values (+ more testing)

+++ include/Margin.hpp
@@ -13,10 +13,10 @@ namespace margin
 	i64 calc_margin(const i32 insta_buy, const i32 insta_sell);
 
 	__attribute__((const, warn_unused_result))
-	i64 calc_profit_with_cut(const i32 margin, const i32 buy_limit, const i32 price_cut);
+	i64 calc_profit_with_cut(const i64 margin, const i32 buy_limit, const i32 price_cut);
 
 	__attribute__((const, warn_unused_result))
-	i64 calc_profit(const i32 buy_price, const i32 sell_price, const i32 buy_limit);
+	i64 calc_profit(const i64 buy_price, const i64 sell_price, const i32 buy_limit);
 
 	__attribute__((const, warn_unused_result))
 	i64 calc_profit(const flips::flip& flip);
